[distributed-systems][2] create homework  - change structure
