+++ 1.Distributed-Systems/master/action/messages/get.py
@@ -1,0 +1,5 @@
+from flask import jsonify
+
+def get_list_action(messages):
+    return jsonify({ "data": messages}), 200
+
